backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gyp==0.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
idna==3.6
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
openai==1.11.1
packaging==23.2
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.1
pydantic_core==2.16.2
PyGObject==3.42.1
PyJWT==2.8.0
pyparsing==2.4.7
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==2.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
systemd-python==234
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
